requirements.txt
pandas
numpy
scikit-learn
scipy
torch
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv
torch-geometric
tqdm
aiohttp
asyncio
transformers
openai
tiktoken
faiss-gpu
h5py
psutil
filelock
matplotlib
